add check for exist member in channel and add get member channels

+++ src/main/java/com/sebapd/chat1b/server/domain/services/ChatChannelService.java
@@ -4,7 +4,8 @@ import com.sebapd.chat1b.server.domain.Channel;
 import com.sebapd.chat1b.server.domain.Member;
 import com.sebapd.chat1b.server.domain.Message;
 import com.sebapd.chat1b.server.domain.exceptions.ChannelNotFoundException;
-import com.sebapd.chat1b.server.domain.exceptions.MemberNotExistInChannel;
+import com.sebapd.chat1b.server.domain.exceptions.MemberAlreadyExistInChannelException;
+import com.sebapd.chat1b.server.domain.exceptions.MemberNotExistInChannelException;
 import com.sebapd.chat1b.server.domain.exceptions.MemberNotFoundException;
 import com.sebapd.chat1b.server.ports.ChannelRepository;
 import com.sebapd.chat1b.server.ports.ChannelService;
@@ -29,6 +30,7 @@ public class ChatChannelService implements ChannelService {
                 .orElseThrow(MemberNotFoundException::new);
         var channel = channelsRepository.getChannelByName(channelName)
                 .orElseThrow(ChannelNotFoundException::new);
+        if(checkIfMemberExistInChannel(channel, member)) throw new MemberAlreadyExistInChannelException();
         channelRepository.addMemberToChannel(member, channel);
     }
 
@@ -36,7 +38,9 @@ public class ChatChannelService implements ChannelService {
     public void removeChannelMember(String memberName, String channelName) {
         var channel = channelsRepository.getChannelByName(channelName)
                 .orElseThrow(ChannelNotFoundException::new);
-        var member = checkIfMemberExistInChannel(channel, memberName);
+        var member = memberRepository.getChatMemberByName(memberName)
+                .orElseThrow(MemberNotFoundException::new);
+        if(!checkIfMemberExistInChannel(channel, member)) throw new MemberNotExistInChannelException();
         channelRepository.removeChannelMember(member, channel);
     }
 
@@ -44,7 +48,9 @@ public class ChatChannelService implements ChannelService {
     public List<Message> getHistory(String channelName, String memberName) {
         var channel = channelsRepository.getChannelByName(channelName)
                 .orElseThrow(ChannelNotFoundException::new);
-        checkIfMemberExistInChannel(channel, memberName);
+        var member = memberRepository.getChatMemberByName(memberName)
+                .orElseThrow(MemberNotFoundException::new);
+        if(!checkIfMemberExistInChannel(channel, member)) throw new MemberNotExistInChannelException();
         var messageList = channel.getMessageList()
                 .stream()
                 .filter(message -> message.getAccessMembersList().contains(memberName))
@@ -53,13 +59,10 @@ public class ChatChannelService implements ChannelService {
         return messageList;
     }
 
-    private Member checkIfMemberExistInChannel(Channel channel, String memberName) {
-        memberRepository.getChatMemberByName(memberName)
-                .orElseThrow(MemberNotFoundException::new);
+    private boolean checkIfMemberExistInChannel(Channel channel, Member member) {
         return channel.getChannelMembers()
                 .stream()
-                .filter(m -> m.getMemberName().equals(memberName))
-                .findAny()
-                .orElseThrow(MemberNotExistInChannel::new);
+                .anyMatch(m -> m.getMemberName().equals(member.getMemberName()));
     }
+
 }
